chore(sandbox): removes dead code
