+++ examples/sandbox/src/lib/theme.ts
@@ -1,0 +1,64 @@
+/** Source: https://github.com/nashamri/spacemacs-theme/blob/master/spacemacs-theme.el */
+export const spacemacs = {
+  "act1": "#e7e5eb",
+  "act2": "#d3d3e7",
+  "base": "#655370",
+  "base-dim": "#a094a2",
+  "bg1": "#fbf8ef",
+  "bg2": "#efeae9",
+  "bg3": "#e3dedd",
+  "bg4": "#d2ceda",
+  "bg-alt": "#efeae9",
+  "border": "#b3b9be",
+  "brown": "#7c3813",
+  "cblk": "#655370",
+  "cblk-bg": "#e8e3f0",
+  "cblk-ln": "#9380b2",
+  "cblk-ln-bg": "#ddd8eb",
+  "cursor": "#100a14",
+  "const": "#4e3163",
+  "comment": "#2aa1ae",
+  "comment-light": "#a49da5",
+  "comment-bg": "#ecf3ec",
+  "comp": "#6c4173",
+  "err": "#e0211d",
+  "func": "#6c3163",
+  "head1": "#3a81c3",
+  "head1-bg": "#edf1ed",
+  "head2": "#2d9574",
+  "head2-bg": "#edf2e9",
+  "head3": "#67b11d",
+  "head3-bg": "#edf2e9",
+  "head4": "#b1951d",
+  "head4-bg": "#f6f1e1",
+  "highlight": "#d3d3e7",
+  "highlight-dim": "#e7e7fc",
+  "keyword": "#a8a8bf",
+  "mat": "#ba2f59",
+  "meta": "#da8b55",
+  "prop": "rgb(107 83 176)",
+  "str": "#2d9574",
+  "suc": "#42ae2c",
+  "ttip": "#8c799f",
+  "ttip-sl": "#c8c6dd",
+  "ttip-bg": "#e2e0ea",
+  "ttype": "#ba2f59",
+  "type": "rgb(143 77 178)",
+  "var": "#715ab1",
+  "war": "#dc752f",
+  "aqua": "#2d9574",
+  "aqua-bg": "#edf2e9",
+  "green": "#67b11d",
+  "green-bg": "#edf2e9",
+  "green-bg-s": "#dae6d0",
+  "cyan": "#21b8c7",
+  "red": "#f2241f",
+  "red-bg": "#faede4",
+  "red-bg-s": "#eed9d2",
+  "blue": "#3a81c3",
+  "blue-bg": "#edf1ed",
+  "blue-bg-s": "#d1dcdf",
+  "magenta": "#a31db1",
+  "yellow": "#b1951d",
+  "yellow-bg": "#f6f1e1",
+} as const
